# api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "私有部署企业协作产品的服务端 API。当前阶段包含管理员登录、管理员创建普通用户、普通用户登录。",
        "title": "AI 原生企业协作服务 API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/admin/settings/info": {
            "get": {
                "description": "管理员读取 App 名称和组织名称。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员设置"
                ],
                "summary": "获取系统基础信息设置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.successEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.infoSettingsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    }
                }
            },
            "put": {
                "description": "管理员更新 App 名称和组织名称。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员设置"
                ],
                "summary": "更新系统基础信息设置",
                "parameters": [
                    {
                        "description": "基础信息设置",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.updateInfoSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.successEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.infoSettingsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/client/conversations/{conversation_id}/messages/images": {
            "post": {
                "description": "普通用户上传 WebP 图片并发送为会话图片消息。图片写入 temporary bucket，消息 body 只保存 file_id。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端消息"
                ],
                "summary": "发送图片消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "conversation_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "客户端消息 ID",
                        "name": "client_message_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "WebP 图片",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.successEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.createMessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.successEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.createMessageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/client/info": {
            "get": {
                "description": "返回客户端启动和登录页展示所需的公开信息，不需要登录。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户端信息"
                ],
                "summary": "获取客户端公开信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.successEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpserver.infoSettingsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpserver.errorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpserver.createMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "$ref": "#/definitions/httpserver.messageResponse"
                }
            }
        },
...
        "httpserver.errorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpserver.errorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httpserver.errorBody"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "httpserver.infoSettingsResponse": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string",
                    "example": "MyGod"
                },
                "authenticated": {
                    "type": "boolean",
                    "example": false
                },
                "organization_name": {
                    "type": "string",
                    "example": "长亭科技"
                },
                "third_party_providers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpserver.publicThirdPartyProviderResponse"
                    }
                }
            }
        },
...
        "httpserver.messageResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object"
                },
                "client_message_id": {
                    "type": "string",
                    "example": "9c08f2dd-0af6-4e99-b486-2f0c841822be"
                },
                "conversation_id": {
                    "type": "string",
                    "example": "7f8d8b84-6d2c-4b12-9a8a-019a7e2787d4"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "id": {
                    "type": "string",
                    "example": "7f8d8b84-6d2c-4b12-9a8a-019a7e2787d4"
                },
                "sender": {
                    "$ref": "#/definitions/httpserver.messageSenderResponse"
                },
                "seq": {
                    "type": "integer",
                    "example": 13
                }
            }
        },
        "httpserver.messageSenderResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "7f8d8b84-6d2c-4b12-9a8a-019a7e2787d4"
                },
                "type": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "httpserver.publicThirdPartyProviderResponse": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string",
                    "example": "company-sso"
                },
                "name": {
                    "type": "string",
                    "example": "企业 SSO"
                }
            }
        },
...
        "httpserver.successEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "httpserver.updateInfoSettingsRequest": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string",
                    "example": "MyGod"
                },
                "organization_name": {
                    "type": "string",
                    "example": "长亭科技"
                }
            }
        },
...
    }
}